+++ Library/Aphid/Core/Application.js
@@ -27,7 +27,7 @@
 
 var Application;
 
-Aphid.Core.Application = Class.create({
+Aphid.Core.Application = Class.create(Aphid.Support.Object, {
 
   displayName: "Aphid.Core.Application",
 
@@ -167,10 +167,6 @@ Aphid.Core.Application = Class.create({
   }
 
 });
-
-// Mixins --------------------------------------------------------------------
-
-Object.extend(Aphid.Core.Application.prototype, Aphid.Support.Properties);
 
 /*
  * Aphid.Core.Application.bootstrap() -> null